Windows port: experimental matmul: fix narrowing in aggregate init
- minimal_matmul: static_cast<uint32_t> num_cores (size_t) in
  shared_variables_t aggregate initializers for both the standard and
  fabric-bound program factories (C2397 narrowing conversion).
- matmul and matmul_decode targets already compiled clean.

+++ ttnn/cpp/ttnn/operations/experimental/minimal_matmul/device/minimal_matmul_fabric_bound_program_factory.cpp
@@ -985,7 +985,7 @@ MinimalMatmulFabricBoundProgramFactory::shared_variables_t minimal_matmul_fabric
     }
 
     return MinimalMatmulFabricBoundProgramFactory::shared_variables_t{
-        num_cores,
+        static_cast<uint32_t>(num_cores),
         cores,
         in0_sender_kernels_id,
         in0_receiver_kernels_id,
